// dcx-ui/packages/dcx-parser/src/model/parameter-mappings.ts
export type ParameterInfo = {
  name: string;
  type: 'bool' | 'enum' | 'number';
  index: number;
  highByteIndex?: number;
  wordOffset?: number;
  wordHighOffset?: number;
  values?: readonly string[];
  unit?: string;
  min?: number;
  max?: number;
  step?: number;
};

export const setupParameters: ParameterInfo[] = [
  {
    name: 'Input Sum Type',
    type: 'enum',
    index: 109,
    values: ['Off', 'A', 'B', 'C', 'A+B', 'A+C', 'B+C'],
  },
  {
    name: 'Input AB Source',
    type: 'enum',
    index: 111,
    values: ['Analog', 'AES/EBU'],
  },
  {
    name: 'Input C Gain',
    type: 'enum',
    index: 113,
    values: ['Line', 'Mic'],
  },
  {
    wordOffset: 5,
    name: 'Output Config',
    type: 'enum',
    index: 115,
    values: ['mono', 'lmhlmh', 'llmmhh', 'lhlhlh'],
  },
  {
    name: 'Stereolink',
    type: 'bool',
    index: 118,
  },
  {
    name: 'Stereolink Mode',
    type: 'enum',
    index: 120,
    values: ['Off', 'A+B', 'A+B+C', 'A+B+C+Sum'],
  },
  {
    name: 'Delay Link',
    type: 'bool',
    index: 122,
  },
  {
    name: 'Crossover Link',
    type: 'bool',
    index: 125,
  },
  {
    name: 'Is Delay Correction On',
    type: 'bool',
    index: 127,
  },
  {
    wordOffset: 10,
    name: 'Air Temperature',
    type: 'number',
    index: 129,
    unit: '°C',
    min: -20,
    max: 50,
    step: 1,
  },
  {
    name: 'Delay Units',
    type: 'enum',
    index: 47,
    values: ['mm', 'inch'],
  },
  {
    name: 'Mute Outs When Powered',
    type: 'bool',
    index: 49,
  },
  {
    wordOffset: 11,
    name: 'Input A Sum Gain',
    type: 'number',
    index: 131,
    highByteIndex: 133,
    unit: 'dB',
    min: -15,
    max: 15,
    step: 0.1,
  },
  {
    wordOffset: 12,
    name: 'Input B Sum Gain',
    type: 'number',
    index: 134,
    highByteIndex: 135,
    unit: 'dB',
    min: -15,
    max: 15,
    step: 0.1,
  },
  {
    wordOffset: 13,
    name: 'Input C Sum Gain',
    type: 'number',
    index: 136,
    highByteIndex: 137,
    unit: 'dB',
    min: -15,
    max: 15,
    step: 0.1,
  },
];